requirements.txt
discord.py>=2.3.0
aiohttp>=3.9.0
aiosqlite>=0.19.0
toml>=0.10.2
